fix assetContainer style, move scroll code
